Support enums declared in other schemas
Previously, I was making the assumption that all enums were declared
as part of the "public" schema, and so any enums in other schemas
were being ignored.

Now, we are accounting for all schemas. Public enums will be dumped as
before, but enums in custom namespaces will be qualified.

Fixes #8